[Misc] Remove patch for fmt, add support fmt version 9, fix #3782

+++ lib/driver/runtimeTool.cpp
@@ -314,7 +314,7 @@ int Tool(struct DriverToolOptions &Opt) noexcept {
           fmt::print("{}\n"sv, (*Result)[I].first.get<double>());
           break;
         case TypeCode::V128:
-          fmt::print("{}\n"sv, (*Result)[I].first.get<uint128_t>());
+          fmt::print("{}\n"sv, uint128((*Result)[I].first.get<uint128_t>()));
           break;
         /// TODO: FuncRef and ExternRef
         default:
